requirements.txt
Flask
pywebview
requests